+++ Robust.Shared/GameObjects/EntitySystemMessages/EntityPausedEvent.cs
@@ -1,14 +1,18 @@
-﻿namespace Robust.Shared.GameObjects
-{
-    public sealed class EntityPausedEvent : EntityEventArgs
-    {
-        public EntityUid Entity { get; }
-        public bool Paused { get; }
+﻿using System;
 
-        public EntityPausedEvent(EntityUid entity, bool paused)
-        {
-            Entity = entity;
-            Paused = paused;
-        }
-    }
+namespace Robust.Shared.GameObjects;
+
+/// <summary>
+/// Raised directed on an entity when it is paused.
+/// </summary>
+[ByRefEvent]
+public readonly record struct EntityPausedEvent;
+
+/// <summary>
+/// Raised directed on an entity when it is unpaused.
+/// </summary>
+[ByRefEvent]
+public readonly record struct EntityUnpausedEvent(TimeSpan PausedTime)
+{
+    public readonly TimeSpan PausedTime = PausedTime;
 }